+++ engine/apps/api/views/shift_swap.py
@@ -4,10 +4,11 @@ from rest_framework import status
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+from rest_framework.serializers import BaseSerializer
 from rest_framework.viewsets import ModelViewSet
 
-from apps.api.permissions import IsOwner, RBACPermission
-from apps.api.serializers.shift_swap import ShiftSwapRequestSerializer
+from apps.api.permissions import AuthenticatedRequest, IsOwner, RBACPermission
+from apps.api.serializers.shift_swap import ShiftSwapRequestListSerializer, ShiftSwapRequestSerializer
 from apps.auth_token.auth import PluginAuthentication
 from apps.mobile_app.auth import MobileAppAuthTokenAuthentication
 from apps.schedules import exceptions
@@ -21,7 +22,7 @@ from common.insight_log import EntityEvent, write_resource_insight_log
 logger = logging.getLogger(__name__)
 
 
-class ShiftSwapViewSet(PublicPrimaryKeyMixin, ModelViewSet):
+class ShiftSwapViewSet(PublicPrimaryKeyMixin[ShiftSwapRequest], ModelViewSet):
     authentication_classes = (MobileAppAuthTokenAuthentication, PluginAuthentication)
     permission_classes = (IsAuthenticated, RBACPermission)
 
@@ -36,7 +37,6 @@ class ShiftSwapViewSet(PublicPrimaryKeyMixin, ModelViewSet):
         "partial_update": [RBACPermission.Permissions.SCHEDULES_WRITE],
         "destroy": [RBACPermission.Permissions.SCHEDULES_WRITE],
         "take": [RBACPermission.Permissions.SCHEDULES_WRITE],
-        "shifts": [RBACPermission.Permissions.SCHEDULES_READ],
     }
 
     is_beneficiary = IsOwner(ownership_field="beneficiary")
@@ -53,19 +53,22 @@ class ShiftSwapViewSet(PublicPrimaryKeyMixin, ModelViewSet):
     serializer_class = ShiftSwapRequestSerializer
     pagination_class = FiftyPageSizePaginator
 
+    def get_serializer_class(self):
+        return ShiftSwapRequestListSerializer if self.action == "list" else super().get_serializer_class()
+
     def get_queryset(self):
         queryset = ShiftSwapRequest.objects.filter(schedule__organization=self.request.auth.organization)
         return self.serializer_class.setup_eager_loading(queryset)
 
-    def perform_destroy(self, instance) -> None:
-        # TODO: should we allow deleting a taken request? if so we will have to undo the overrides that were generated
+    def perform_destroy(self, instance: ShiftSwapRequest) -> None:
+        # TODO: should we allow deleting a taken request?
 
         super().perform_destroy(instance)
         write_resource_insight_log(instance=instance, author=self.request.user, event=EntityEvent.DELETED)
 
         update_shift_swap_request_message.apply_async((instance.pk,))
 
-    def perform_create(self, serializer) -> None:
+    def perform_create(self, serializer: BaseSerializer[ShiftSwapRequest]) -> None:
         beneficiary = self.request.user
         shift_swap_request = serializer.save(beneficiary=beneficiary)
 
@@ -73,7 +76,7 @@ class ShiftSwapViewSet(PublicPrimaryKeyMixin, ModelViewSet):
 
         create_shift_swap_request_message.apply_async((shift_swap_request.pk,))
 
-    def perform_update(self, serializer) -> None:
+    def perform_update(self, serializer: BaseSerializer[ShiftSwapRequest]) -> None:
         prev_state = serializer.instance.insight_logs_serialized
         serializer.save()
         shift_swap_request = serializer.instance
@@ -88,15 +91,8 @@ class ShiftSwapViewSet(PublicPrimaryKeyMixin, ModelViewSet):
 
         update_shift_swap_request_message.apply_async((shift_swap_request.pk,))
 
-    @action(methods=["get"], detail=True)
-    def shifts(self, request, pk) -> Response:
-        shift_swap = self.get_object()
-        result = {"events": shift_swap.shifts()}
-
-        return Response(result, status=status.HTTP_200_OK)
-
     @action(methods=["post"], detail=True)
-    def take(self, request, pk) -> Response:
+    def take(self, request: AuthenticatedRequest, pk: str) -> Response:
         shift_swap = self.get_object()
 
         try: