+++ engine/apps/schedules/models/on_call_schedule.py
@@ -51,6 +51,7 @@ if typing.TYPE_CHECKING:
 
     from apps.alerts.models import EscalationPolicy
     from apps.auth_token.models import ScheduleExportAuthToken
+    from apps.schedules.models import ShiftSwapRequest
     from apps.slack.models import SlackUserGroup
     from apps.user_management.models import Organization, Team
 
@@ -165,7 +166,9 @@ class OnCallScheduleQuerySet(PolymorphicQuerySet):
 
 
 class OnCallSchedule(PolymorphicModel):
+    custom_shifts: "RelatedManager['CustomOnCallShift']"
     organization: "Organization"
+    shift_swap_requests: "RelatedManager['ShiftSwapRequest']"
     slack_user_group: typing.Optional["SlackUserGroup"]
     team: typing.Optional["Team"]
 
@@ -331,13 +334,13 @@ class OnCallSchedule(PolymorphicModel):
 
     def filter_events(
         self,
-        datetime_start,
-        datetime_end,
-        with_empty=False,
-        with_gap=False,
-        filter_by=None,
-        all_day_datetime=False,
-        from_cached_final=False,
+        datetime_start: datetime.datetime,
+        datetime_end: datetime.datetime,
+        with_empty: bool = False,
+        with_gap: bool = False,
+        filter_by: str | None = None,
+        all_day_datetime: bool = False,
+        from_cached_final: bool = False,
     ) -> ScheduleEvents:
         """Return filtered events from schedule."""
         shifts = (
@@ -396,13 +399,15 @@ class OnCallSchedule(PolymorphicModel):
 
         return events
 
-    def final_events(self, datetime_start, datetime_end):
+    def final_events(self, datetime_start: datetime.datetime, datetime_end: datetime.datetime) -> ScheduleEvents:
         """Return schedule final events, after resolving shifts and overrides."""
         events = self.filter_events(datetime_start, datetime_end, with_empty=True, with_gap=True, all_day_datetime=True)
         events = self._resolve_schedule(events, datetime_start, datetime_end)
         return events
 
-    def filter_swap_requests(self, datetime_start, datetime_end):
+    def filter_swap_requests(
+        self, datetime_start: datetime.datetime, datetime_end: datetime.time
+    ) -> "RelatedManager['ShiftSwapRequest']":
         swap_requests = self.shift_swap_requests.filter(  # starting before but ongoing
             swap_start__lt=datetime_start, swap_end__gte=datetime_start
         ).union(
@@ -627,12 +632,14 @@ class OnCallSchedule(PolymorphicModel):
             "overloaded_users": overloaded_users,
         }
 
-    def _apply_swap_requests(self, events, datetime_start, datetime_end) -> ScheduleEvents:
+    def _apply_swap_requests(
+        self, events: ScheduleEvents, datetime_start: datetime.datetime, datetime_end: datetime.datetime
+    ) -> ScheduleEvents:
         """Apply swap requests details to schedule events."""
         # get swaps requests affecting this schedule / time range
         swaps = self.filter_swap_requests(datetime_start, datetime_end)
 
-        def _insert_event(index, event):
+        def _insert_event(index: int, event: ScheduleEvent) -> int:
             # add event, if any, to events list in the specified index
             # return incremented index if the event was added
             if event is None:
@@ -642,11 +649,14 @@ class OnCallSchedule(PolymorphicModel):
 
         # apply swaps sequentially
         for swap in swaps:
+            if swap.is_past_due:
+                # ignore untaken expired requests
+                continue
             i = 0
             while i < len(events):
                 event = events.pop(i)
 
-                if event["start"] > swap.swap_end or event["end"] < swap.swap_start:
+                if event["start"] >= swap.swap_end or event["end"] <= swap.swap_start:
                     # event outside the swap period, keep as it is and continue
                     i = _insert_event(i, event)
                     continue