+++ engine/apps/schedules/tests/test_shift_swap_request.py
@@ -2,11 +2,14 @@ import datetime
 from unittest.mock import patch
 
 import pytest
+import pytz
 from django.utils import timezone
 
 from apps.schedules import exceptions
 from apps.schedules.models import CustomOnCallShift, ShiftSwapRequest
 from apps.user_management.models import User
+
+ROTATION_START = datetime.datetime(2150, 8, 29, 0, 0, 0, 0, tzinfo=pytz.UTC)
 
 
 @pytest.mark.django_db
@@ -120,7 +123,6 @@ def test_take_own_ssr(shift_swap_request_setup) -> None:
         ssr.take(beneficiary)
 
 
-@pytest.mark.skip(reason="Skipping to unblock release")
 @pytest.mark.django_db
 def test_related_shifts(shift_swap_request_setup, make_on_call_shift) -> None:
     ssr, beneficiary, _ = shift_swap_request_setup()